lib/libcamera-source: Add a read() call to the event handler
In the libcamera-source we are spoofing fd POLLIN events that are
available with V4L2 devnodes through the use of a pipe, the write
end of which is written to in the requestComplete callback. This
needs an equivalent read() call in the event handler callback to
ensure the fd isn't continuously triggering POLLIN on select.

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>