lib: Guard against buffer overflow in .fill_buffer()
In the .fill_buffer() op we copy the entire image that was given
as a source to the sink buffer. There exists the possibility of users
defining formats in configfs whos buffer size would be smaller than
the size of the image, which would lead to a buffer overflow during
the memcpy() - guard against that by copying at maximum the size of
the sink buffer.

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>