lib: configfs: Move uvc_formats to dedicated file
UVC Gadget can support more formats than the MJPEG and YUYV that we
currently list; move the uvc_formats array and the definitions of
the UVC GUIDs to a separate file to keep things neat in configfs.c
and list some extra supported formats.

The listed ones are those which I could easily verify support for
via vivid, but in principle any uncompressed format ought to work.

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>